id_file_search/requirements.txt
requests
beautifulsoup4
lxml
google-genai
python-dotenv
anthropic
